refactor: extract matrix.thread_projection from thread_membership (ARCH-B-3 step 2)
